feat(publish-action): only when package.json version is updated, the action will start
